fix: test if secret is nil before trying to use it (#65)
* fix: test if secret is nil before trying to use it

* chore: add Alex to code owners

* fix: check nil on testsuite secret delete as well

+++ client/testsuites/v1/testsuites.go
@@ -157,7 +157,7 @@ func (s TestSuitesClient) Delete(name string) error {
 		return err
 	}
 
-	if secretExists {
+	if secretExists && secret != nil {
 		return s.Client.Delete(context.Background(), secret)
 	}
 	return nil
